Move test_types into test_sampling
